% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{alaskaAirlinesFlights}
\alias{alaskaAirlinesFlights}
\title{Example data - Alaska Airlines Flights}
\format{An object of class \code{data.frame} with 714 rows and 20 columns.}
\usage{
alaskaAirlinesFlights
}
\description{
Example data - Alaska Airlines Flights
}
\references{
\url{http://users.stat.umn.edu/~almquist/3811_examples/all_alaska_flights.csv}
}
\author{
Alaska Airlines
}
\keyword{datasets}
