% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{ggwrapper}
\alias{ggwrapper}
\title{GGPlot Wrapper}
\usage{
ggwrapper(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
ggplot2
}
\description{
Computes the mean, variance and sd of a vector
}
\examples{
data(d)
ggwrapper(d)
}
