% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizFunctions.R
\name{matx}
\alias{matx}
\title{Computation of a Matrix and Vector}
\usage{
matx(mat, vec)
}
\arguments{
\item{mat}{A Matrix}

\item{vec}{A Vector}
}
\value{
Scalar
}
\description{
Computes (x)(matrix)(x^T)
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p1.rda"))
matx(a, x)

}
