% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{Q1P3}
\alias{Q1P3}
\title{Example data - Quiz 1 Problem 3}
\format{An object of class \code{data.frame} with 5 rows and 5 columns.}
\usage{
Q1P3
}
\description{
Example data - Quiz 1 Problem 3
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q1p3.txt
}
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
