% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizFunctions.R
\name{std}
\alias{std}
\title{Standardize Matrix}
\usage{
std(dat)
}
\arguments{
\item{dat}{A 2-D Matrix with numerical values}
}
\value{
A 2-D Matrix
}
\description{
Column standarization of a matrix
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p1.rda"))
std(a)

}
